Refactored theme event to its own module apart from the config service
