+++ src/app/components/app-body.component.ts
@@ -1,7 +1,8 @@
-import {customElement, property} from 'lit/decorators.js';
-import {css, html, LitElement, nothing, unsafeCSS} from 'lit';
+import { customElement, property } from 'lit/decorators.js';
+import { css, html, LitElement, nothing, unsafeCSS } from 'lit';
 import Fontawesome from '../utilities/fontawesome';
-import {ConfigService, ThemeEvent} from '../services/config.service';
+import { ConfigService } from '../services/config.service';
+import { ThemeEvent } from '../events/theme-event';
 
 const AppBodyComponentStyles = css`
     :root {
